python/requirements.txt
mido==1.3.2
numpy==1.26.4
packaging==23.2
pyFluidSynth==1.3.3
python-rtmidi==1.5.8